lsproto: consolidate codec tests into lsp_json_test.go
Move the round-trip, strictness, and UnmarshalParams tests out of their own
files and into lsp_json_test.go alongside the existing JSON (un)marshal
tests, where they are more discoverable.
